# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventApp API",
        "description": "API REST pour la gestion d'événements collaboratifs - Plateforme de création, gestion et inscription à des événements",
        "version": "1.0.0",
        "contact": {
            "name": "EventApp Support",
            "email": "support@eventapp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Liste des \u00e9v\u00e9nements",
                "operationId": "get_api_events_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "isFree",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des \u00e9v\u00e9nements"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Cr\u00e9er un \u00e9v\u00e9nement",
                "operationId": "post_api_events_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "startDate",
                                    "location",
                                    "maxParticipants"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Conf\u00e9rence Tech 2024"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Une conf\u00e9rence sur les derni\u00e8res technologies"
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-15T14:00:00Z"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-15T18:00:00Z"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Paris, France"
                                    },
                                    "maxParticipants": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.99
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "tech",
                                            "conference"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "\u00c9v\u00e9nement cr\u00e9\u00e9 avec succ\u00e8s",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Conf\u00e9rence Tech 2024"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Une conf\u00e9rence sur les derni\u00e8res technologies"
                                                },
                                                "startDate": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-15T14:00:00Z"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Paris, France"
                                                },
                                                "maxParticipants": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "isPublished": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Donn\u00e9es invalides"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
...
        "/api/events/{id}/publish": {
            "patch": {
                "tags": [
                    "Events"
                ],
                "summary": "Publier un \u00e9v\u00e9nement",
                "operationId": "patch_api_events_publish",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u00c9v\u00e9nement publi\u00e9"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    },
                    "404": {
                        "description": "\u00c9v\u00e9nement non trouv\u00e9"
                    }
                }
            }
        },
        "/api/events/popular": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "\u00c9v\u00e9nements populaires",
                "operationId": "get_api_events_popular",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u00c9v\u00e9nements populaires"
                    }
                }
            }
        },
        "/api/events/upcoming": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "\u00c9v\u00e9nements \u00e0 venir",
                "operationId": "get_api_events_upcoming",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u00c9v\u00e9nements \u00e0 venir"
                    }
                }
            }
        },
...
        "/api/registrations/events/{id}": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "S'inscrire \u00e0 un \u00e9v\u00e9nement",
                "operationId": "post_api_registrations_register",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Inscription cr\u00e9\u00e9e"
                    },
                    "400": {
                        "description": "Inscription impossible"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    },
                    "404": {
                        "description": "\u00c9v\u00e9nement non trouv\u00e9"
                    }
                }
            }
        },
...
        "/api/users/me/events": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Mes \u00e9v\u00e9nements organis\u00e9s",
                "operationId": "get_api_users_my_events",
                "responses": {
                    "200": {
                        "description": "\u00c9v\u00e9nements organis\u00e9s"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
...
    },
...
}